working Vite, TailwindCSS on pong game (basic frontend framework has been established
