+++ frontend/index.html
@@ -1,13 +1,60 @@
-<!DOCTYPE html>
-<html>
-	<head>
-		<meta charset="UTF-8">
-		<meta name="viewport", content="width=device-width, initial-scale=1.0">
-		<title>Pong</title>
-		<link rel="stylesheet" href="pong.css">
-		<script defer src="dist/pong.js"></script>
-	</head>
-	<body>
-		<canvas id="board"></canvas>
-	</body>
+<!doctype html>
+<html lang="en" class="dark">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Pong</title>
+
+  </head>
+  <body class="min-h-screen bg-white dark:bg-black text-black dark:text-white">
+    <main class="max-w-6xl mx-auto p-4">
+      <div class="grid grid-cols-1 md:grid-cols-4 gap-6">
+      <!-- Main game area (3/4 on medium+ screen and full width on mobile) -->
+      <section class="md:col-span-3 md:col-start-1">
+        <div id="pong-container" class="flex justify-center">
+          <canvas
+            id="board"
+            class="block rounded bg-gray-800 w-full max-w-[800px] aspect-[4/3]">
+          </canvas>
+        </div>
+      </section>
+          
+      <!-- Sidebar: controls 1/4 on medium and full width on mobile 
+       THIS SHOULD HAVE THE BUTTON ON THE RIGHT BUT IT IS APPEARING UNDER-->
+      <aside class="order-first md:order-none md:col-span-1">
+        <div class="sticky top-4 space-y-3">
+          <button
+            id="toggleDark"
+            class="ml-auto px-4 py-2 text-sm font-medium bg-blue-100 text-blue-900 rounded-md">
+            Toggle Dark Mode
+          </button>
+        </div>
+        <!--PUT OTHER BUTTONS HERE IF WE NEED THEM OR A MENU -->
+        <div id="error-message" class="hidden text-red-600 text-center"> </div>
+      </aside>
+    </div>
+  </main>
+  <script type="module" src="/src/main.ts"></script>
+  <script>
+    document.addEventListener("DOMContentLoaded", () => {
+      const btn = document.getElementById('toggleDark')
+
+      function updateButtonText() {
+        btn.textContent = document.documentElement.classList.contains('dark') ? 'Light Mode' : 'Dark Mode';
+      }
+      
+      if (btn) {
+        //set initial label
+        updateButtonText();
+        btn.addEventListener('click', () => {
+          document.documentElement.classList.toggle('dark');
+          btn.setAttribute('aria-pressed',
+            String(document.documentElement.classList.contains('dark'))
+          );
+          updateButtonText();
+        });
+      }
+    });  
+    </script>
+  </body>
 </html>
