--- frontend/pong.css
@@ -1,9 +1,0 @@
-body {
-	text-align: center;	
-}
-
-#board {
-	background-color: black;
-	border-top: 5px solid pink;
-	border-bottom: 5px solid pink;
-}
