+++ frontend/postcss.config.ts
@@ -1,0 +1,6 @@
+// postcss.config.ts
+export default {
+  plugins: {
+    '@tailwindcss/postcss': {},
+  },
+}
