+++ frontend/tsconfig.json
@@ -1,46 +1,25 @@
 {
-  // Visit https://aka.ms/tsconfig to read more about this file
   "compilerOptions": {
-    // File Layout
-    "rootDir": "src",
-    "outDir": "dist",
-
-    // Environment Settings
-    // See also https://aka.ms/tsconfig/module
+    "target": "ES2022",
+    "useDefineForClassFields": true,
     "module": "ESNext",
-    "target": "ES6",
-    //"types": [],
-    // For nodejs:
-     "lib": ["ES6", "DOM"],
-    // "types": ["node"],
-    // and npm install -D @types/node
-
-    // Other Outputs
-    "sourceMap": true,
-    //"declaration": true,
-    //"declarationMap": true,
-
-    // Stricter Typechecking Options
-    //"noUncheckedIndexedAccess": true,
-    //"exactOptionalPropertyTypes": true,
-
-    // Style Options
-    // "noImplicitReturns": true,
-    // "noImplicitOverride": true,
-    // "noUnusedLocals": true,
-    // "noUnusedParameters": true,
-    // "noFallthroughCasesInSwitch": true,
-    // "noPropertyAccessFromIndexSignature": true,
-
-    // Recommended Options
-    "strict": true,
-    //"jsx": "react-jsx",
-    //"verbatimModuleSyntax": true,
-    //"isolatedModules": true,
-    //"noUncheckedSideEffectImports": true,
-    //"moduleDetection": "force",
+    "lib": ["ES2022", "DOM", "DOM.Iterable"],
     "skipLibCheck": true,
-    "noEmitOnError": true
+
+    /* Bundler mode */
+    "moduleResolution": "bundler",
+    "allowImportingTsExtensions": true,
+    "verbatimModuleSyntax": true,
+    "moduleDetection": "force",
+    "noEmit": true,
+
+    /* Linting */
+    "strict": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "erasableSyntaxOnly": true,
+    "noFallthroughCasesInSwitch": true,
+    "noUncheckedSideEffectImports": true
   },
-  "include": ["src/**/*"]
+  "include": ["src"]
 }
