--- package.json
@@ -1,7 +1,0 @@
-{
-  "devDependencies": {
-    "@types/jest": "^30.0.0",
-    "jest": "^30.0.5",
-    "ts-jest": "^29.4.1"
-  }
-}
